move more out of tdma
