+++ must-hop/src/policy/tdma/slots.rs
@@ -1,0 +1,98 @@
+use core::fmt;
+
+#[cfg(not(feature = "in_std"))]
+use defmt::debug;
+use heapless::Vec;
+#[cfg(feature = "in_std")]
+use log::debug;
+use serde::{Deserialize, Serialize};
+
+pub struct SlotMask {
+    mask: u8,
+}
+impl Default for SlotMask {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl fmt::Debug for SlotMask {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "Taken Slots: [")?;
+        let mut first = true;
+
+        for i in 0..32 {
+            if self.is_taken(i) {
+                if !first {
+                    write!(f, ", ")?;
+                }
+                write!(f, "{}", i)?;
+                first = false;
+            }
+        }
+        write!(f, "]")
+    }
+}
+
+#[cfg(not(feature = "in_std"))]
+impl defmt::Format for SlotMask {
+    fn format(&self, fmt: defmt::Formatter) {
+        defmt::write!(fmt, "Taken Slots: [");
+        let mut first = true;
+
+        for i in 0..32 {
+            if self.is_taken(i) {
+                if !first {
+                    defmt::write!(fmt, ", ");
+                }
+                // We use {=u8} to tell defmt exactly what type it is sending over the wire
+                defmt::write!(fmt, "{=u8}", i);
+                first = false;
+            }
+        }
+        defmt::write!(fmt, "]");
+    }
+}
+
+impl SlotMask {
+    pub const fn new() -> Self {
+        Self { mask: 0 }
+    }
+    /// To set a slot inside mask
+    pub fn claim(&mut self, slot: u8) {
+        // shift 1 over to slot pos, and or with mask
+        self.mask |= 1 << slot;
+    }
+
+    /// Check given slot is occupied
+    pub fn is_taken(&self, slot: u8) -> bool {
+        (self.mask & (1 << slot)) != 0
+    }
+
+    pub fn as_u32(&self) -> u8 {
+        self.mask
+    }
+
+    /// Get the next available slot given another node's mask and yours. Uses the node_id to avoid conflicts in race conditions
+    pub fn slot_assignment_strat(
+        &self,
+        max_slots: u8,
+        another_mask: u8,
+        node_id: u8,
+    ) -> Option<u8> {
+        let combined_mask = SlotMask {
+            mask: self.mask | another_mask,
+        };
+        let start_offset = node_id % max_slots;
+
+        (0..max_slots).find_map(|i| {
+            let slot = (start_offset + i) % max_slots;
+            debug!("looking in slot {}", slot);
+            if !combined_mask.is_taken(slot) {
+                Some(slot)
+            } else {
+                None
+            }
+        })
+    }
+}
